refactor(layout): Use Astro components for everything
Since there are performance benefits to using Astro's static components
where possible, we'll use them everywhere unless we explicitly need
client side interactivity. We're retaining the Svelte dependency for
that.
