+++ src/components/comments/UtterancesComments.astro
@@ -1,0 +1,18 @@
+<div>
+	<script
+		src="https://utteranc.es/client.js"
+		repo="icorbrey/isaaccorbrey.com"
+		crossorigin="anonymous"
+		issue-term="og:title"
+		theme="github-dark"
+		label="comments"
+		is:inline
+		async
+	></script>
+</div>
+
+<style>
+	div {
+		margin-top: 2rem;
+	}
+</style>
