+++ src/components/tags/Tag.astro
@@ -1,0 +1,54 @@
+---
+import { classes } from "../../utils/classes";
+
+type Props = {
+	description?: string;
+	large: boolean;
+	href?: string;
+};
+
+const {
+	description = "This tag does not yet have a description.",
+	large = false,
+	href,
+} = Astro.props;
+---
+
+{
+	!!href ? (
+		<a {href} title={description} class={classes({ large })}>
+			#<slot />
+		</a>
+	) : (
+		<span class={classes({ large })}>
+			#<slot />
+		</span>
+	)
+}
+
+<style>
+	a,
+	span {
+		padding: 0.125rem 0.5rem;
+		border-radius: 1rem;
+
+		text-transform: uppercase;
+		letter-spacing: 0.0625rem;
+		font-size: 0.875rem;
+		font-weight: 500;
+
+		border: solid 0.1rem var(--foreground-4);
+		color: var(--foreground-3);
+	}
+
+	a:hover {
+		color: var(--foreground-2);
+	}
+
+	.large {
+		font-size: 2.5rem;
+
+		padding: 0.25rem 1.25rem;
+		border-radius: 3rem;
+	}
+</style>
