--- src/controls/Comments.svelte
@@ -1,20 +1,0 @@
-{#if true}
-	<div>
-		<script
-			src="https://utteranc.es/client.js"
-			repo="icorbrey/isaaccorbrey.com"
-			crossorigin="anonymous"
-			issue-term="og:title"
-			theme="github-dark"
-			label="comments"
-			async
-		>
-		</script>
-	</div>
-{/if}
-
-<style>
-	div {
-		margin-top: 2rem;
-	}
-</style>
