+++ src/layouts/Root.astro
@@ -1,6 +1,8 @@
 ---
+import ScreenColumn from '../components/controls/ScreenColumn.astro';
+import Footer from '../components/controls/Footer.astro';
+import Header from '../components/controls/Header.astro';
 import { articleSchema } from "../utils/openGraph";
-import Layout from "../controls/Layout.svelte";
 import { z } from "astro:content";
 import "../style/global.css";
 
@@ -68,8 +70,10 @@ const siteName = "Isaac Corbrey";
 		}
 	</head>
 	<body>
-		<Layout>
+		<ScreenColumn>
+			<Header />
 			<slot />
-		</Layout>
+			<Footer />
+		</ScreenColumn>
 	</body>
 </html>
