+++ src/pages/404.mdx
@@ -1,4 +1,4 @@
-import Hero from "../controls/Hero.svelte";
+import Hero from "../components/controls/Hero.astro";
 import Root from "../layouts/Root.astro";
 
 <Root title="Page not found">
