+++ src/pages/index.mdx
@@ -1,10 +1,10 @@
-import ProjectList from "../features/projects/ProjectList.svelte";
-import NoteList from "../features/notes/NoteList.svelte";
-import ButtonLink from "../controls/ButtonLink.svelte";
-import { noteStore } from "../features/notes/store";
-import Hero from "../controls/Hero.svelte";
+import ProjectList from "../components/projects/ProjectList.astro";
+import ButtonLink from "../components/controls/ButtonLink.astro";
+import NoteList from "../components/notes/NoteList.astro";
+import Hero from "../components/controls/Hero.astro";
+import Row from "../components/controls/Row.astro";
+import { noteStore } from "../data/notes";
 import Root from "../layouts/Root.astro";
-import Row from "../controls/Row.svelte";
 
 <Root
 	description={
